requirements.txt
streamlit
streamlit_authenticator
pandas 
google-generativeai 
pathlib  
ipython 
matplotlib
python-dotenv
wordcloud